Split visitor into its own module
